+++ tests/NpgsqlFSharpAnalyzer.Tests/Tests.fs
@@ -32,7 +32,7 @@ let tests =
             | None -> failwith "Could not crack project"
             | Some context ->
                 let operationBlocks = SyntacticAnalysis.findSqlOperations context
-                Expect.equal 4 (List.length operationBlocks) "Found four operation blocks"
+                Expect.equal 7 (List.length operationBlocks) "Found four operation blocks"
         }
 
         test "Syntactic Analysis: reading queries with [<Literal>] query" {