+++ tests/examples/hashing/AssemblyInfo.fs
@@ -4,20 +4,20 @@ open System.Reflection
 
 [<assembly: AssemblyTitleAttribute("examples")>]
 [<assembly: AssemblyProductAttribute("NpgsqlFSharpAnalyzer")>]
-[<assembly: AssemblyVersionAttribute("3.3.0")>]
-[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-04-17T00:00:00.0000000")>]
-[<assembly: AssemblyFileVersionAttribute("3.3.0")>]
-[<assembly: AssemblyInformationalVersionAttribute("3.3.0")>]
+[<assembly: AssemblyVersionAttribute("3.4.0")>]
+[<assembly: AssemblyMetadataAttribute("ReleaseDate","2020-05-19T00:00:00.0000000")>]
+[<assembly: AssemblyFileVersionAttribute("3.4.0")>]
+[<assembly: AssemblyInformationalVersionAttribute("3.4.0")>]
 [<assembly: AssemblyMetadataAttribute("ReleaseChannel","release")>]
-[<assembly: AssemblyMetadataAttribute("GitHash","ca8053aec45712b7f48789d2026ec86364ab57fe")>]
+[<assembly: AssemblyMetadataAttribute("GitHash","179f9fa515334affde1f14aea077328215d27a70")>]
 do ()
 
 module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "examples"
     let [<Literal>] AssemblyProduct = "NpgsqlFSharpAnalyzer"
-    let [<Literal>] AssemblyVersion = "3.3.0"
-    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-04-17T00:00:00.0000000"
-    let [<Literal>] AssemblyFileVersion = "3.3.0"
-    let [<Literal>] AssemblyInformationalVersion = "3.3.0"
+    let [<Literal>] AssemblyVersion = "3.4.0"
+    let [<Literal>] AssemblyMetadata_ReleaseDate = "2020-05-19T00:00:00.0000000"
+    let [<Literal>] AssemblyFileVersion = "3.4.0"
+    let [<Literal>] AssemblyInformationalVersion = "3.4.0"
     let [<Literal>] AssemblyMetadata_ReleaseChannel = "release"
-    let [<Literal>] AssemblyMetadata_GitHash = "ca8053aec45712b7f48789d2026ec86364ab57fe"
+    let [<Literal>] AssemblyMetadata_GitHash = "179f9fa515334affde1f14aea077328215d27a70"